restructure file to support autoload
