+++ composer.json
@@ -16,7 +16,7 @@
     },
     "autoload": {
         "psr-4": {
-            "shukriyusof\\kolonialisme\\": ""
+            "shukriyusof\\kolonialisme\\": "src/"
         }
     }
 }
