+++ src/School.php
@@ -1,119 +1,6 @@
 <?php
 namespace shukriyusof\kolonialisme;
 
-class State {
-
-    public static function getAllStates(){
-        $listingStates = file_get_contents(__DIR__.'/inc/state.json');
-        $listingStates = json_decode($listingStates,true);
-        
-        return $listingStates;
-    }
-
-    public static function getStateByShortname($params):array {
-        $listingStates = file_get_contents(__DIR__.'/inc/state.json');
-        $listingStates = json_decode($listingStates,true);
-
-        foreach ($listingStates as $state) {
-            if(!empty($params) && ($state['shortname'] == $params)){
-                return $state;
-            }
-        }
-
-        return $state;
-    }
-
-    public static function getStateByName($params):array {
-        $listingStates = file_get_contents(__DIR__.'/inc/state.json');
-        $listingStates = json_decode($listingStates,true);
-
-        foreach ($listingStates as $state) {
-            if(!empty($params) && ($state['name'] == $params)){
-                return $state;
-            }
-        }
-
-        return $state;
-    }
-
-    public static function getStateById($params):array {
-        $listingStates = file_get_contents(__DIR__.'/inc/state.json');
-        $listingStates = json_decode($listingStates,true);
-
-        foreach ($listingStates as $state) {
-            if(!empty($params) && $state['id'] == $params){
-                    return $state;
-            }
-        }
-
-        return $state;
-    }
-
-    public static function getStateByIso($params):array {
-        $listingStates = file_get_contents(__DIR__.'/inc/state.json');
-        $listingStates = json_decode($listingStates,true);
-
-        foreach ($listingStates as $state) {
-            if(!empty($params) && ($state['iso'] == $params)){
-                    return $state;
-            }
-        }
-
-        return $state;
-    }
-}
-
-class District {
-    public static function getAllDistricts(){
-        $listingDistricts = file_get_contents(__DIR__.'/inc/district.json');
-        $listingDistricts = json_decode($listingDistricts,true);
-        
-        return $listingDistricts;
-    }
-
-    // 2)getDistrictById
-    public static function getDistrictById($params):array {
-        $listingDistricts = file_get_contents(__DIR__.'/inc/district.json');
-        $listingDistricts = json_decode($listingDistricts,true);
-
-        foreach ($listingDistricts as $district) {
-            if(!empty($params) && ($params ==  $district['id'])){
-                return $district;
-            }
-        }
-
-        return $district;
-    }
-
-    // 3)getDistrictByName
-    public static function getDistrictByName($params):array {
-        $listingDistricts = file_get_contents(__DIR__.'/inc/district.json');
-        $listingDistricts = json_decode($listingDistricts,true);
-
-        foreach ($listingDistricts as $district) {
-            if(!empty($params) && ($params ==  $district['name'])){
-                return $district;
-            }
-        }
-
-        return $district;
-    }
-
-    // 4)getDistrictByStateId
-    public static function getDistrictByStateId($params):array {
-        $listingDistricts = file_get_contents(__DIR__.'/inc/district.json');
-        $listingDistricts = json_decode($listingDistricts,true);
-
-        foreach ($listingDistricts as $district) {
-            if(!empty($params) && ($params ==  $district['state_id'])){
-                return $district;
-            }
-        }
-
-        return $district;
-    }
-}
-
 class School {
     // return all schools without any requirement
     public static function getAllSchools(){
